+++ admin/app/Global.scala
@@ -1,10 +1,10 @@
 import common.CloudWatchApplicationMetrics
 import conf.Management
-import model.AdminLifecycle
+import model.{PorterLifecycle, AdminLifecycle}
 import play.api.mvc.{Results, RequestHeader}
 import scala.concurrent.Future
 
-object Global extends AdminLifecycle with CloudWatchApplicationMetrics with Results {
+object Global extends AdminLifecycle with PorterLifecycle with CloudWatchApplicationMetrics with Results {
   override lazy val applicationName = Management.applicationName
 
   override def onError(request: RequestHeader, ex: Throwable) = Future.successful(InternalServerError(
