+++ admin/app/services/CloudWatch.scala
@@ -2,7 +2,7 @@ package services
 
 import com.amazonaws.services.cloudwatch.AmazonCloudWatchAsyncClient
 import com.amazonaws.handlers.AsyncHandler
-import conf.{PorterHealthCheckPage, Configuration}
+import conf.{AdminHealthCheckPage, Configuration}
 import com.amazonaws.services.cloudwatch.model._
 import scala.collection.JavaConversions._
 import common.Logging
@@ -75,7 +75,7 @@ trait CloudWatch extends implicits.Futures {
       exception match {
         // temporary till JVM bug fix comes out
         // see https://blogs.oracle.com/joew/entry/jdk_7u45_aws_issue_123
-        case e: Exception if e.getMessage.contains("JAXP00010001") => PorterHealthCheckPage.setUnhealthy()
+        case e: Exception if e.getMessage.contains("JAXP00010001") => AdminHealthCheckPage.setUnhealthy()
         case _ =>
       }
     }
