--- porter/app/controllers/IndexController.scala
@@ -1,7 +1,0 @@
-package controllers
-
-import play.api.mvc.{ Action, Controller }
-
-object IndexController extends Controller {
-  def index() = Action { Ok("Ok") }
-}
