--- porter/conf/application.conf
@@ -1,47 +1,0 @@
-# This is the main configuration file for the application.
-# ~~~~~
-
-application: {
-    # The secret key is used to secure cryptographics functions.
-    # If you deploy your application to several instances be sure to use the same key!
-    secret: ${APP_SECRET}
-    langs: "en"
-}
-
-logger: {
-    # Even though we configure logback using conf/logger.xml, Play still inherits a standard logback configuration
-    # which defines 'play' and 'application' loggers which create `logs/application.log`. The default configuration
-    # also logs ROOT at ERROR which is not useful for us.
-    root: INFO,
-    play: INFO,
-    application: DEBUG
-}
-
-
-############################################################
-#
-# Threadpool config
-# see http://www.playframework.com/documentation/2.2.x/ThreadPools
-#
-############################################################
-
-play {
-  akka {
-    akka.loggers = ["akka.event.Logging$DefaultLogger", "akka.event.slf4j.Slf4jLogger"]
-    loglevel = WARNING
-    actor {
-      default-dispatcher = {
-        fork-join-executor {
-          parallelism-factor = 1.0
-          parallelism-max = 24
-        }
-      }
-      java-futures = {
-        fork-join-executor {
-          parallelism-factor = 1.0
-          parallelism-max = 1
-        }
-      }
-    }
-  }
-}
