--- porter/conf/deploy.json
@@ -1,28 +1,0 @@
-{
-    "packages":{
-        "frontend-porter":{
-            "type":"autoscaling",
-            "apps":[ "frontend::porter" ],
-            "data":{
-                "secondsToWait":1200,
-                "port":18080,
-                "healthcheckGrace":20,
-                "bucket":"aws-frontend-artifacts",
-                "healthcheck_paths":[
-                    "/management/healthcheck"
-                ]
-            }
-        }
-    },
-    "recipes":{
-        "default":{
-            "depends" : ["artifactUpload", "deploy"]
-        },
-        "deploy":{
-            "actionsBeforeApp": ["frontend-porter.deploy"]
-        },
-        "artifactUpload":{
-            "actionsBeforeApp": ["frontend-porter.uploadArtifacts"]
-        }
-    }
-}
