--- porter/conf/play.plugins
@@ -1,1 +1,0 @@
-1000:com.gu.management.play.InternalManagementPlugin
